Fix log color loss and LogLocation not working
